# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.0",
    "title": "Muscufit API documentation",
    "description": "Documentation of Muscufit API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brewenlpg@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8181/Muscufit/api/v1",
  "servers": [
    {
      "url": "http://localhost:8181/Muscufit/api/v1/"
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "authorization",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "API endpoint to get user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User id"
          },
          {
            "name": "filters",
            "in": "query",
            "description": "definition of filters",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Model of user response",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "API Endpoint to add new user",
        "parameters": [
          {
            "name": "newUser",
            "in": "body",
            "description": "User informations",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered with success !!!"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "adress": {
          "type": "string",
          "example": "50 rue des mouettes"
        }
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "about": {
          "type": "string",
          "example": ""
        }
      },
      "required": [
        "name",
        "age"
      ]
    },
...
  }
}